Refactored example application
