+++ integration_test_suite/jmx_example_application/src/main/java/io/prometheus/jmx/OptionalValue.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) The Prometheus jmx_exporter Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.prometheus.jmx;
+
+import java.lang.management.ManagementFactory;
+import java.util.Optional;
+import javax.management.MBeanServer;
+import javax.management.ObjectName;
+
+/** Class to implement OptionalValue */
+public class OptionalValue implements OptionalValueMBean {
+
+    /** Constructor */
+    public OptionalValue() {
+        // INTENTIONALLY BLANK
+    }
+
+    @Override
+    public Optional<Integer> getValue() {
+        return Optional.of(345);
+    }
+
+    /**
+     * Method to register the MBean
+     *
+     * @throws Exception If an error occurs during registration
+     */
+    public void register() throws Exception {
+        MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
+        mBeanServer.registerMBean(
+                new OptionalValue(), new ObjectName("io.prometheus.jmx:type=optionalValue"));
+    }
+}
