+++ integration_test_suite/jmx_example_application/src/main/java/io/prometheus/jmx/PerformanceMetricsMBean.java
@@ -16,14 +16,8 @@
 
 package io.prometheus.jmx;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
 import javax.management.openmbean.CompositeData;
-import javax.management.openmbean.CompositeDataSupport;
-import javax.management.openmbean.CompositeType;
 import javax.management.openmbean.OpenDataException;
-import javax.management.openmbean.OpenType;
-import javax.management.openmbean.SimpleType;
 
 /** Interface to implement PerformanceMetricsMBean */
 public interface PerformanceMetricsMBean {
@@ -35,42 +29,4 @@ public interface PerformanceMetricsMBean {
      * @throws OpenDataException OpenDataException
      */
     CompositeData getPerformanceMetrics() throws OpenDataException;
-}
-
-/** Class to implement PerformanceMetrics */
-class PerformanceMetrics implements PerformanceMetricsMBean {
-
-    private final CompositeData compositeData;
-
-    /**
-     * Constructor
-     *
-     * @throws OpenDataException OpenDataException
-     */
-    public PerformanceMetrics() throws OpenDataException {
-        compositeData = build();
-    }
-
-    @Override
-    public CompositeData getPerformanceMetrics() {
-        return compositeData;
-    }
-
-    private CompositeData build() throws OpenDataException {
-        Map<String, Number> data = new LinkedHashMap<>();
-
-        data.put("ActiveSessions", 2L);
-        data.put("Bootstraps", 4L);
-        data.put("BootstrapsDeferred", 6L);
-
-        String[] names = {"ActiveSessions", "Bootstraps", "BootstrapsDeferred"};
-        String[] descriptions = {"ActiveSessions", "Bootstraps", "BootstrapsDeferred"};
-        OpenType<?>[] types = {SimpleType.LONG, SimpleType.LONG, SimpleType.LONG};
-
-        CompositeType compositeType =
-                new CompositeType(
-                        "PerformanceMetrics", "Example composite data", names, descriptions, types);
-
-        return new CompositeDataSupport(compositeType, data);
-    }
 }
