+++ integration_test_suite/jmx_example_application/src/main/java/io/prometheus/jmx/StringValueMBean.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package io.prometheus.jmx;
 
 /** Class to implement StringValueMBean */
@@ -24,18 +25,4 @@ public interface StringValueMBean {
      * @return text
      */
     String getText();
-}
-
-/** Class to implement CustomValue */
-class StringValue implements StringValueMBean {
-
-    /** Constructor */
-    public StringValue() {
-        // INTENTIONALLY BLANK
-    }
-
-    @Override
-    public String getText() {
-        return "value";
-    }
 }
